refactor!: Merge `arguments.hpp` into `log.hpp`
